Replace Sidebar store with events/emits/props
Toggling the SideBar's 'visible' prop to false, or initializing it as false,
results in the CoreUI Sidebar component emitting a 'hide' event (a callback fired after hiding).
Sometimes, we want to obey this 'hide' emit: specifically, when the user, on
mobile, taps another part of the page. But if the 'hide' emit is received as a result of
the user directly requesting the sidebar to be hidden, we should NOT then try to toggle the
sidebar a second time, which would undo the user's action.