Refactor defineEventHandler wrapping
This was to avoid having a handler fed into a another defineEventHandler which
would wrap the first EventHandler in a second EventHandler. Although the docs pointed
us in the direction of doing such a double-wrapping, we found this way to avoid doing so.

A downside of this is that endpoints are forced to use a specific function (e.g.
'defineEventHandler' or in this case 'defineCachedEventHandler') since they all route through
the wrapper's choice of function - this means if we wanted to use e.g. defineLazyEventHandler
from h3 in some endpoint, it (probably?) couldn't use our custom wrapper, unless we introduce some
switch statement that lets endpoints specify which EventHandler wrapping function to use. I am aware
of the following wrapping function choices:
- defineEventHandler
- defineCachedEventHandler
- defineLazyEventHandler