Stop using CoreUI CFormSelect component; use native HTML
We found the documentation and behavior of the CFormSelect component too
confusing to be worth any benefit. In particular, it did
not seem to honour the initial v-model value, so we had to manually
set the 'selected' attribute for each option (which results in a
recalculation of options every time one is selected) in order to get the
correct default value to be honoured - and even then, it was confusingly
slow or flakey about it.