Pass dedupe 'defer' option to useFetch
https://nuxt.com/docs/api/composables/use-fetch\#params

On slow networks we saw that about half the time, a status poll
request resulted in useFetch cacheing the status as 'running' for a long
time (minutes) even though new requests were returning the correct
status ('completed'). Given that the dedupe option fixes it, it was
presumably because new requests were constantly being cancelled when
the polling interval came around again, before they had had a chance
to return anything (on a slow network speed).