Re-implement resetting tooltips/crosshairs on (un)synchronised charts
Previously, we followed the lead of a highcharts demo in replacing the 'reset' function
of the Pointer prototype with 'undefined' in order to keep tooltips and crosshairs
visible when the mouse leaves a chart that we're trying to synchronise
-  https://www.highcharts.com/demo/highcharts/synchronized-charts

However, this has the undesirable effect of preventing unrelated charts (such as
our cost pie or, now, cost column chart) behaving normally, since the prototype
is shared by all Highcharts charts. We had to intervene to manually reset other
charts, or write styles and apply classes to hide the tooltips manually for those
unrelated other charts.

A better solution, since it allows other charts to behave as expected by default,
is to use Highcharts' 'wrap' function for extending private functions to wrap the
Pointer.onContainerMouseLeave function. In the wrapper, we check if the chart that
is triggering onContainerMouseLeave is among our collection of synchronised charts,
and prevent the wrapped function from executing if so; else we allow the original
function to proceed. And we manually trigger the Pointer.reset() for all charts
in the collection of synchronised charts when the mouse leaves the collection
container - since onContainerMouseLeave had otherwise been responsible for triggering
Pointer.reset().