Provide nuxt context explicitly when refreshing statuses server-side
See https://nuxt.com/docs/4.x/api/composables/use-nuxt-app\#runwithcontext

This is to prevent an error (really a warning, since everything continued to work as expected)
that was thrown whenever a comparison page was loaded on the server side. The error message is
reproduced below. I infer that it was caused by awaited calls to refreshScenarioStatus that did
not have access to the nuxt context (as the docs say above, nuxt/vue drop context when doing awaits,
in order to prevent shared state between different users' requests during SSR). I didn't want to
solve the problem by switching to using dollar-fetch instead of useFetch, as that would mean
preventing the calls from ever running on the server side (that's the point of using useFetch).
So we manage the context explicitly, allowing it to be provided for the pinia method.

ERROR  [unhandledRejection] [nuxt] A composable that requires access to the Nuxt instance was called outside of a plugin, Nuxt hook, Nuxt middleware, or Vue setup function. This is probably not a Nuxt bug. Find out more at https://nuxt.com/docs/guide/concepts/auto-imports\#vue-and-nuxt-composables.

    at useNuxtApp (node_modules/nuxt/dist/app/nuxt.js:251:13)
    at useAsyncData (node_modules/nuxt/dist/app/composables/asyncData.js:30:55)
    at useFetch (node_modules/nuxt/dist/app/composables/fetch.js:57:59)
    at Proxy.refreshScenarioStatus (stores/appStore.ts:164:58)
    at Proxy.wrappedAction (node_modules/pinia/dist/pinia.mjs:1083:18)
    at stores/appStore.ts:230:20
    at wrappedFn (node_modules/@vue/reactivity/dist/reactivity.cjs.js:861:19)
    at Array.map (<anonymous>)
    at apply (node_modules/@vue/reactivity/dist/reactivity.cjs.js:869:27)
    at Proxy.map (node_modules/@vue/reactivity/dist/reactivity.cjs.js:793:12)