don't change id of chart container els
We're needlessly changing the id of the container elements for time series charts when they are toggled to the other time series in a group. This could potentially confuse Highcharts. So I made the container element names constant by basing them on the series group index instead of the series itself.