Update CapacitiesAndInterventionsLegend props
I've now updated this component to take a boolean prop 'comparisonMode'. This is preferable to deriving the mode from the store's state, because otherwise, during navigation from a comparison to a single scenario, this legend distractingly flashes some changes. That happens because, during the navigation, the store state is updated before the new page components get rendered.